Compression and Comparrison
I have created 2 subroutines that we may be able to use.  The first one which I have called "combineAnswer" takes the values present in the player's answer list (which I have called "playerAnswer" for convenience, but this can easily be changed if it conflicts with the actual variable name we come up with) and condenses them into a single string. Bassically, ["1","0","1"] becomes "101". This allows us to compare the the player's answer to the answer that is present in the key or the dictionary containing our Riddles (which for testing purposes I named "answer" and only listed one key and value.) The only issue is that the code will not accurately compare a list to a string, but it will compare two strings within lists, so I re-inserted the answer_str as a single value in a list. Thus "101" becomes ["101"] and that allows for the second subroutine (which I have named "compareAnswers") merely compares the two lists and determins if they are equal are not. This subroutine should be heavily modified in the future.