Full Code -- First Draft
I went ahead and tried to put together a full code of what we need to have before 5/5/17. It appears to be pretty sloppy and my Pi died on me last night, so I am unable to test it until later today. Let me know if there are any glaring issues with it.